requirements.txt
docx==0.2.4
langchain==1.0.3
langchain_huggingface==1.0.0
langchain_openai==1.0.1
pinecone==7.3.0
python-dotenv==1.2.1
streamlit==1.51.0
transformers==4.57.1
